+++ rpcs3/Emu/RSX/Common/texture_cache_utils.h
@@ -105,36 +105,6 @@ namespace rsx
 		constexpr operator enum_type() const { return cause; }
 	};
 
-	struct typeless_xfer
-	{
-		bool src_is_typeless = false;
-		bool dst_is_typeless = false;
-		bool src_is_depth = false;
-		bool dst_is_depth = false;
-		u32 src_gcm_format = 0;
-		u32 dst_gcm_format = 0;
-		f32 src_scaling_hint = 1.f;
-		f32 dst_scaling_hint = 1.f;
-		texture_upload_context src_context = texture_upload_context::blit_engine_src;
-		texture_upload_context dst_context = texture_upload_context::blit_engine_dst;
-
-		void analyse()
-		{
-			if (src_is_typeless && dst_is_typeless)
-			{
-				if (src_scaling_hint == dst_scaling_hint &&
-					src_scaling_hint != 1.f)
-				{
-					if (src_is_depth == dst_is_depth)
-					{
-						src_is_typeless = dst_is_typeless = false;
-						src_scaling_hint = dst_scaling_hint = 1.f;
-					}
-				}
-			}
-		}
-	};
-
 
 
 	/**
