+++ rpcs3/Emu/RSX/VK/VKTextureCache.h
@@ -1097,130 +1097,7 @@ namespace vk
 
 		bool blit(rsx::blit_src_info& src, rsx::blit_dst_info& dst, bool interpolate, rsx::vk_render_targets& m_rtts, vk::command_buffer& cmd)
 		{
-			struct blit_helper
-			{
-				vk::command_buffer* commands;
-				VkFormat format;
-				blit_helper(vk::command_buffer *c) : commands(c) {}
-
-				void scale_image(vk::image* src, vk::image* dst, areai src_area, areai dst_area, bool interpolate, bool /*is_depth*/, const rsx::typeless_xfer& xfer_info)
-				{
-					const auto src_aspect = vk::get_aspect_flags(src->info.format);
-					const auto dst_aspect = vk::get_aspect_flags(dst->info.format);
-
-					vk::image* real_src = src;
-					vk::image* real_dst = dst;
-
-					if (xfer_info.src_is_typeless)
-					{
-						auto internal_width = src->width() * xfer_info.src_scaling_hint;
-						auto format = vk::get_compatible_sampler_format(vk::get_current_renderer()->get_formats_support(), xfer_info.src_gcm_format);
-
-						// Transfer bits from src to typeless src
-						real_src = vk::get_typeless_helper(format);
-						src_area.x1 = (u16)(src_area.x1 * xfer_info.src_scaling_hint);
-						src_area.x2 = (u16)(src_area.x2 * xfer_info.src_scaling_hint);
-
-						vk::copy_image_typeless(*commands, src, real_src, { 0, 0, (s32)src->width(), (s32)src->height() }, { 0, 0, (s32)internal_width, (s32)src->height() }, 1,
-							vk::get_aspect_flags(src->info.format), vk::get_aspect_flags(format));
-					}
-
-					if (xfer_info.dst_is_typeless)
-					{
-						auto internal_width = dst->width() * xfer_info.dst_scaling_hint;
-						auto format = vk::get_compatible_sampler_format(vk::get_current_renderer()->get_formats_support(), xfer_info.dst_gcm_format);
-
-						// Transfer bits from dst to typeless dst
-						real_dst = vk::get_typeless_helper(format);
-						dst_area.x1 = (u16)(dst_area.x1 * xfer_info.dst_scaling_hint);
-						dst_area.x2 = (u16)(dst_area.x2 * xfer_info.dst_scaling_hint);
-
-						vk::copy_image_typeless(*commands, dst, real_dst, { 0, 0, (s32)dst->width(), (s32)dst->height() }, { 0, 0, (s32)internal_width, (s32)dst->height() }, 1,
-							vk::get_aspect_flags(dst->info.format), vk::get_aspect_flags(format));
-					}
-					else if (xfer_info.dst_context == rsx::texture_upload_context::framebuffer_storage)
-					{
-						if (xfer_info.src_context != rsx::texture_upload_context::blit_engine_dst &&
-							xfer_info.src_context != rsx::texture_upload_context::framebuffer_storage)
-						{
-							// Data moving to rendertarget, where byte ordering has to be preserved
-							// NOTE: This is a workaround, true accuracy would require all RTT<->cache transfers to invoke this step but thats too slow
-							// Sampling is ok; image view swizzle will work around it
-							if (dst->info.format == VK_FORMAT_B8G8R8A8_UNORM)
-							{
-								// For this specific format, channel ordering is faked via custom remap, undo this before transfer
-								VkBufferImageCopy copy{};
-								copy.imageExtent = src->info.extent;
-								copy.imageOffset = { 0, 0, 0 };
-								copy.imageSubresource = { VK_IMAGE_ASPECT_COLOR_BIT, 0, 0, 1 };
-
-								const auto scratch_buf = vk::get_scratch_buffer();
-								const auto data_length = src->info.extent.width * src->info.extent.height * 4;
-
-								const auto current_layout = src->current_layout;
-								vk::change_image_layout(*commands, real_src, VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL, { VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1 });
-								vkCmdCopyImageToBuffer(*commands, src->value, src->current_layout, scratch_buf->value, 1, &copy);
-								vk::change_image_layout(*commands, real_src, current_layout, { VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1 });
-
-								vk::insert_buffer_memory_barrier(*commands, scratch_buf->value, 0, data_length,
-									VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT,
-									VK_ACCESS_TRANSFER_WRITE_BIT, VK_ACCESS_SHADER_READ_BIT);
-
-								vk::get_compute_task<vk::cs_shuffle_32>()->run(*commands, scratch_buf, data_length);
-
-								vk::insert_buffer_memory_barrier(*commands, scratch_buf->value, 0, data_length,
-									VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT,
-									VK_ACCESS_SHADER_WRITE_BIT, VK_ACCESS_TRANSFER_READ_BIT);
-
-								real_src = vk::get_typeless_helper(src->info.format);
-								vk::change_image_layout(*commands, real_src, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, {VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1});
-
-								vkCmdCopyBufferToImage(*commands, scratch_buf->value, real_src->value, real_src->current_layout, 1, &copy);
-							}
-						}
-					}
-
-					// Checks
-					if (src_area.x2 <= src_area.x1 || src_area.y2 <= src_area.y1 || dst_area.x2 <= dst_area.x1 || dst_area.y2 <= dst_area.y1)
-					{
-						LOG_ERROR(RSX, "Blit request consists of an empty region descriptor!");
-						return;
-					}
-
-					if (src_area.x1 < 0 || src_area.x2 >(s32)real_src->width() || src_area.y1 < 0 || src_area.y2 >(s32)real_src->height())
-					{
-						LOG_ERROR(RSX, "Blit request denied because the source region does not fit!");
-						return;
-					}
-
-					if (dst_area.x1 < 0 || dst_area.x2 >(s32)real_dst->width() || dst_area.y1 < 0 || dst_area.y2 >(s32)real_dst->height())
-					{
-						LOG_ERROR(RSX, "Blit request denied because the destination region does not fit!");
-						return;
-					}
-
-					const auto src_width = src_area.x2 - src_area.x1;
-					const auto src_height = src_area.y2 - src_area.y1;
-					const auto dst_width = dst_area.x2 - dst_area.x1;
-					const auto dst_height = dst_area.y2 - dst_area.y1;
-
-					copy_scaled_image(*commands, real_src->value, real_dst->value, real_src->current_layout, real_dst->current_layout, src_area.x1, src_area.y1, src_width, src_height,
-						dst_area.x1, dst_area.y1, dst_width, dst_height, 1, dst_aspect, real_src->info.format == real_dst->info.format,
-						interpolate ? VK_FILTER_LINEAR : VK_FILTER_NEAREST, real_src->info.format, real_dst->info.format);
-
-					if (real_dst != dst)
-					{
-						auto internal_width = dst->width() * xfer_info.dst_scaling_hint;
-						vk::copy_image_typeless(*commands, real_dst, dst, { 0, 0, (s32)internal_width, (s32)dst->height() }, { 0, 0, (s32)dst->width(), (s32)dst->height() }, 1,
-							vk::get_aspect_flags(real_dst->info.format), vk::get_aspect_flags(dst->info.format));
-					}
-
-					change_image_layout(*commands, dst, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, {(VkImageAspectFlags)dst_aspect, 0, dst->info.mipLevels, 0, dst->info.arrayLayers});
-					format = dst->info.format;
-				}
-			}
-			helper(&cmd);
-
+			blitter helper(&cmd);
 			auto reply = upload_scaled_image(src, dst, interpolate, cmd, m_rtts, helper, cmd, const_cast<const VkQueue>(m_submit_queue));
 
 			if (reply.succeeded)
